// main/dashboard/src/shared/types/voxlab/material-types.ts
export interface MaterialSettings {
    tint: string;
    opacity: number;
    metalness: number;
    roughness: number;
    emissiveColor: string;
    emissiveIntensity: number;
    flatShading: boolean;
    clearcoat: number;
    clearcoatRoughness: number;
    ior: number;
    sheen: number;
    sheenColor: string;
    anisotropy: number;
}